// src/shared/enums/segment-identification.enums.ts
export enum SegmentIdentification {
  _1 = "Prefeituras",
  _2 = "Saneamento",
  _3 = "Energia Elétrica e Gás",
  _4 = "Telecomunicações",
  _5 = "Órgãos Governamentais",
  _6 = "Carnes e Assemelhados ou demais Empresas / Órgãos que serão identificadas através do CNPJ",
  _7 = "Multas de trânsito",
  _9 = "Uso exclusivo do banco",
}